moved helper such as ContainerHelper & HealthHelper into a different directory
