+++ src/main/java/teamport/aether/mixin/container/ContainerInventoryMixinZanite.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import teamport.aether.accessory.api.ICountArmor;
+import teamport.aether.api.ICountArmor;
 import teamport.aether.items.AetherArmorMaterial;
 
 @Mixin(value = ContainerInventory.class, remap = false)
