+++ src/main/java/teamport/aether/mixin/entity/player/LifeShardExtraHealth.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import teamport.aether.accessory.api.IVariableHealthPlayer;
+import teamport.aether.api.IVariableHealthPlayer;
 
 @Mixin(value = Player.class, remap = false)
 public class LifeShardExtraHealth
